[BACKLOG-5405] - Split UT/IT for project: pentaho-reporting-engine-classic-core-platform-plugin - Renamed classes to reflect the nature of the tests in it; - Separated one of the classes that contained both unit tests and integration tests;
