+++ test-src/org/pentaho/reporting/platform/plugin/PageableHTMLIT.java
@@ -21,7 +21,6 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.util.HashMap;
-import java.util.Map;
 
 import junit.framework.TestCase;
 import org.pentaho.platform.api.engine.IPentahoSession;
@@ -29,13 +28,7 @@ import org.pentaho.platform.engine.core.system.PentahoSessionHolder;
 import org.pentaho.platform.engine.core.system.StandaloneSession;
 import org.pentaho.test.platform.engine.core.MicroPlatform;
 
-/**
- * Unit tests for the ReportingComponent.
- * 
- * @author Michael D'Amour
- */
-public class ReportingComponentTest extends TestCase {
-
+public class PageableHTMLIT extends TestCase {
   private MicroPlatform microPlatform;
   private File tmp;
 
@@ -46,6 +39,7 @@ public class ReportingComponentTest extends TestCase {
 
     microPlatform = MicroPlatformFactory.create();
     microPlatform.start();
+
     IPentahoSession session = new StandaloneSession();
     PentahoSessionHolder.setSession( session );
   }
@@ -55,43 +49,7 @@ public class ReportingComponentTest extends TestCase {
     microPlatform.stop();
   }
 
-  public void testReportDefinitionAsInput() throws Exception {
-    // create an instance of the component
-    SimpleReportingComponent rc = new SimpleReportingComponent();
-    // create/set the InputStream
-    FileInputStream reportDefinition = new FileInputStream( "./resource/solution/test/reporting/report.prpt" ); //$NON-NLS-1$
-    Map<String, Object> inputs = new HashMap<String, Object>();
-    inputs.put( SimpleReportingComponent.REPORT_DEFINITION_INPUT, reportDefinition );
-    rc.setInputs( inputs );
-    rc.setOutputType( "application/pdf" ); //$NON-NLS-1$
-
-    FileOutputStream outputStream = new FileOutputStream( new File(tmp, System.currentTimeMillis() + ".pdf" )); //$NON-NLS-1$ //$NON-NLS-2$
-    rc.setOutputStream( outputStream );
-
-    // validate the component
-    assertTrue( rc.validate() );
-    // execute the component
-    assertTrue( rc.execute() );
-  }
-
-  public void testPDF() throws Exception {
-    // create an instance of the component
-    SimpleReportingComponent rc = new SimpleReportingComponent();
-    // create/set the InputStream
-    FileInputStream reportDefinition = new FileInputStream( "./resource/solution/test/reporting/report.prpt" ); //$NON-NLS-1$
-    rc.setReportDefinitionInputStream( reportDefinition );
-    rc.setOutputType( "application/pdf" ); //$NON-NLS-1$
-
-    FileOutputStream outputStream = new FileOutputStream( new File(tmp, System.currentTimeMillis() + ".pdf" )); //$NON-NLS-1$ //$NON-NLS-2$
-    rc.setOutputStream( outputStream );
-
-    // validate the component
-    assertTrue( rc.validate() );
-    // execute the component
-    assertTrue( rc.execute() );
-  }
-
-  public void testHTML() throws Exception {
+  public void testPageCount() throws Exception {
     // create an instance of the component
     SimpleReportingComponent rc = new SimpleReportingComponent();
     // create/set the InputStream
@@ -99,13 +57,61 @@ public class ReportingComponentTest extends TestCase {
     rc.setReportDefinitionInputStream( reportDefinition );
     rc.setOutputType( "text/html" ); //$NON-NLS-1$
 
+    // turn on pagination, by way of input (typical mode for xaction)
+    HashMap<String, Object> inputs = new HashMap<String, Object>();
+    inputs.put( "paginate", "true" ); //$NON-NLS-1$ //$NON-NLS-2$
+    inputs.put( "accepted-page", "0" ); //$NON-NLS-1$ //$NON-NLS-2$
+    rc.setInputs( inputs );
+
     FileOutputStream outputStream = new FileOutputStream( new File(tmp, System.currentTimeMillis() + ".html" )); //$NON-NLS-1$ //$NON-NLS-2$
     rc.setOutputStream( outputStream );
 
-    // validate the component
-    assertTrue( rc.validate() );
     // execute the component
     assertTrue( rc.execute() );
+
+    // make sure this report has 8 pages (we know this report will produce 8 pages with sample data)
+    assertEquals( 8, rc.getPageCount() );
+  }
+
+  public void testPaginatedHTML() throws Exception {
+    // create an instance of the component
+    SimpleReportingComponent rc = new SimpleReportingComponent();
+    // create/set the InputStream
+    FileInputStream reportDefinition = new FileInputStream( "resource/solution/test/reporting/report.prpt" ); //$NON-NLS-1$
+    rc.setReportDefinitionInputStream( reportDefinition );
+    rc.setOutputType( "text/html" ); //$NON-NLS-1$
+
+    // turn on pagination
+    rc.setPaginateOutput( true );
+    assertTrue( rc.isPaginateOutput() );
+
+    // turn it back off
+    rc.setPaginateOutput( false );
+    assertFalse( rc.isPaginateOutput() );
+
+    // turn on pagination, by way of input (typical mode for xaction)
+    HashMap<String, Object> inputs = new HashMap<String, Object>();
+    inputs.put( "paginate", "true" ); //$NON-NLS-1$ //$NON-NLS-2$
+    inputs.put( "accepted-page", "0" ); //$NON-NLS-1$ //$NON-NLS-2$
+    rc.setInputs( inputs );
+
+    FileOutputStream outputStream = new FileOutputStream( new File(tmp, System.currentTimeMillis() + ".html" )); //$NON-NLS-1$ //$NON-NLS-2$
+    rc.setOutputStream( outputStream );
+
+    // check the accepted page
+    assertEquals( 0, rc.getAcceptedPage() );
+
+    // make sure pagination is really on
+    assertTrue( rc.isPaginateOutput() );
+    // validate the component
+    assertTrue( rc.validate() );
+
+    // execute the component
+    assertTrue( rc.execute() );
+
+    // make sure this report has 8 pages (we know this report will produce 8 pages with sample data)
+    assertEquals( 8, rc.getPageCount() );
+
   }
 
 }
